replace all version checks for DUNE_ISTL by ones for DUNE_COMMON
this is a work around for the build system bug that DUNE_ISTL_VERSION*
macros are not generated under some circumstances if dunecontrol is
used. (For some reason, the HAVE_DUNE_ISTL macro is correctly set,
though.) Since issue gets *really* annoying for me because it reliably
breaks my build and I've got to hand-edit config.h and since it is
unlikely that dune-istl and dune-common exhibit a different version,
we can also check for the version of the latter.

a proper build system fix is appreciated.