Do not copy the wellSolution() member from previous state.
Instead, we let this member always be initialized from the other data
members. In addition to simplifying the code, this is necessary for
perfect restarting. Before this, the restarted state would be
initialized from the other data members while the not-restarted
state (at the same timestep) would be copied from the previous state.
This caused floating-point roundoff differences that would eventually
creep into all state variables. Although insignificant, it would prevent
bitwise-perfect restarting.