Hopefully fixes issue with network file systems and async_output.
There have been reports by people that parallel runs with async_output=true
that write to network file systems do not behave as expected.

This commit is a guess on how to fix these issues. I cannot test it myself.
Judging from the code the only thing that could go wrong before is that the
program might have terminated before all write processes finished. This was
possible because the writer thread was detached when setting up the output
writer and we never checked whether everything was actually written before
the program exists.

With this commit, we do not detach anymore. Instead we signal the thread that no
objects will be written anymore and join it at the end of the run method.