Introduce more proper error handling to asynchronous output writing.
The dispatch method now checks whether there was any error, previously.
In this case it clears the queue. In any case a tuple with the error
indicator and the messages is returned which is used outside for error
handling. For error handling for the last output write step will be done
in the ThreadHandleQueue destructor that is run by the main thread.