Extend 'Wells' Structure to Preserve 'Kh' per Connection
This is mostly to assist the creation of the SCON restart array.

In particular, the helper function

    WellManagerDetail::computeWellIndex()

now returns a structure that holds both the connection
transmissibility factor and the effective Kh product for one
particular connection.  Those values are then passed on to function

    add_well()

as one additional parameter, which we store in the usual way in the
new Wells::conn_Kh data member.

Update the callers of add_well() accordingly.