SPE1 ebos test: also compare the UNRST files
this segfaulted due to a small pecularity of the compareECL
utility. thanks to [at]akva2 for pointing this out.

Also, this updates the reference solutions for this test in an attempt
to fix the jenkins complaint: The test seems to be quite sensitive to
the used compiler flags and Jenkins is not able to update the
reference solutions yet.